refactor: move JWT and middleware utility logic to shared sdk-core package
