+++ packages/sdk-core/src/index.ts
@@ -1,3 +1,8 @@
+export type { NextgenSession, AuthState, UnauthState, AuthResult, NextgenMiddlewareOptions } from "./types.js";
+export { HOP_BY_HOP, INTERNAL_HEADERS, upgradeSessionCookie, matchesRoutes, filterResponseHeaders } from "./middleware-utils.js";
+export { verifyJwt, decodeJwt, base64UrlDecode, JWKS_TTL_MS } from "./jwt.js";
+export type { JwtPayload, JwtHeader, DecodedJwt, VerifyJwtOptions } from "./jwt.js";
+
 export type ZitadelEnvironment = "development" | "preview" | "production";
 
 export type ZitadelRuntime = {
